switch back to individual log messages for now

+++ events/message_delete.go
@@ -144,9 +144,18 @@ func (bot *Bot) messageDelete(m *gateway.MessageDeleteEvent) {
 		Timestamp: discord.NewTimestamp(msg.MsgID.Time()),
 	}
 
-	client := webhook.New(wh.ID, wh.Token)
-	bot.Queue(wh, "message_delete", client, e)
-	bot.DB.DeleteMessage(msg.MsgID)
+	_, err = webhook.New(wh.ID, wh.Token).ExecuteAndWait(webhook.ExecuteData{
+		AvatarURL: bot.Router.Bot.AvatarURL(),
+		Embeds:    []discord.Embed{e},
+	})
+	if err == nil {
+		bot.DB.DeleteMessage(msg.MsgID)
+	} else {
+		bot.DB.Report(db.ErrorContext{
+			Event:   "message_delete",
+			GuildID: m.GuildID,
+		}, err)
+	}
 }
 
 func hasAnyPrefixLower(s string, prefixes ...string) bool {
